style: reformat Nix and Lua files for consistency
This commit applies consistent formatting and indentation across
multiple Nix and Lua configuration files. The changes include
reordering function arguments, aligning brackets, and improving
readability without altering logic or functionality.

+++ nixos-config/modules/shared/whmconfig.nix
@@ -1,5 +1,9 @@
-{lib, config, pkgs, ...}:
-let
+{
+  lib,
+  config,
+  pkgs,
+  ...
+}: let
   inherit (lib) mkOption mkEnableOption types mkIf;
   cfg = config.home.whmConfig;
   homeDir = config.home.homeDirectory;
@@ -71,12 +75,15 @@ in {
         fi
       '';
 
-      whm2Link = let 
-        getLink = opts: let enabled = cfg.link."${opts.name}"; in [
-          ( { inherit enabled; } // opts )
+      whm2Link = let
+        getLink = opts: let
+          enabled = cfg.link."${opts.name}";
+        in [
+          ({inherit enabled;} // opts)
         ];
-        linkList = ([] ++
-          (getLink {
+        linkList = (
+          []
+          ++ (getLink {
             name = "nvim";
             files = [
               {
@@ -84,8 +91,8 @@ in {
                 dst = "${homeDir}/.config/nvim";
               }
             ];
-          }) ++
-          (getLink {
+          })
+          ++ (getLink {
             name = "vim";
             files = [
               {
@@ -93,8 +100,8 @@ in {
                 dst = "${homeDir}/.vimrc";
               }
             ];
-          }) ++
-          (getLink {
+          })
+          ++ (getLink {
             name = "tmux";
             files = [
               {
@@ -102,8 +109,8 @@ in {
                 dst = "${homeDir}/.tmux.conf";
               }
             ];
-          }) ++
-          (getLink {
+          })
+          ++ (getLink {
             name = "wezterm";
             files = [
               {
@@ -117,47 +124,49 @@ in {
             ];
           })
         );
-    
-        in
-        (builtins.concatStringsSep "\n" ([
+      in (builtins.concatStringsSep "\n" [
         ''
           if [[ $_WHMCONFIG_INSTALLED == "true" ]]; then
             true;
-        '' 
-        (builtins.concatStringsSep "\n" (builtins.map (prog: (builtins.concatStringsSep "\n"
-          (builtins.map (link:
-            let 
-              name = builtins.baseNameOf link.dst;
-            in if prog.enabled then ''
-            if [[ -e "${link.dst}" ]]; then
-              if [[ ! -L "${link.dst}" ]]; then
-                echo "[!!] ${name} configuration already exists. Create Backup..." >&2;
-                extension=".backup-$(date +%Y-%m-%d_%H-%M-%S)";
-                mv "${link.dst}" "${link.dst}.$extension";
-                echo "[!!] Backup created: ${link.dst}.$extension" >&2;
-              elif [[ "$(readlink '${link.dst}')" != "${link.src}" ]]; then
-                echo "[!!] Remove link $(stat -c '%N' '${link.dst}')" >&2;
-                unlink "${link.dst}";
-              fi;
-            fi;
-            if [[ ! -e "${link.dst}" ]]; then
-              echo "[*] Link ${link.src} -> ${link.dst}" >&2;
-              ln -s "${link.src}" "${link.dst}";
-              echo "[+] ${name} configuration linked." >&2;
-            fi
-          '' else ''
-            if [[ -L "${link.dst}" && $(readlink "${link.dst}") == "${link.src}" ]]; then
-              echo "[*] Remove link $(stat -c '%N' '${link.dst}')" >&2;
-              unlink "${link.dst}";
-            fi;
-          '') prog.files))
-        ) linkList))
+        ''
+        (builtins.concatStringsSep "\n" (builtins.map (
+            prog: (builtins.concatStringsSep "\n"
+              (builtins.map (link: let
+                name = builtins.baseNameOf link.dst;
+              in
+                if prog.enabled
+                then ''
+                  if [[ -e "${link.dst}" ]]; then
+                    if [[ ! -L "${link.dst}" ]]; then
+                      echo "[!!] ${name} configuration already exists. Create Backup..." >&2;
+                      extension=".backup-$(date +%Y-%m-%d_%H-%M-%S)";
+                      mv "${link.dst}" "${link.dst}.$extension";
+                      echo "[!!] Backup created: ${link.dst}.$extension" >&2;
+                    elif [[ "$(readlink '${link.dst}')" != "${link.src}" ]]; then
+                      echo "[!!] Remove link $(stat -c '%N' '${link.dst}')" >&2;
+                      unlink "${link.dst}";
+                    fi;
+                  fi;
+                  if [[ ! -e "${link.dst}" ]]; then
+                    echo "[*] Link ${link.src} -> ${link.dst}" >&2;
+                    ln -s "${link.src}" "${link.dst}";
+                    echo "[+] ${name} configuration linked." >&2;
+                  fi
+                ''
+                else ''
+                  if [[ -L "${link.dst}" && $(readlink "${link.dst}") == "${link.src}" ]]; then
+                    echo "[*] Remove link $(stat -c '%N' '${link.dst}')" >&2;
+                    unlink "${link.dst}";
+                  fi;
+                '') prog.files))
+          )
+          linkList))
         ''
           else
             echo "[-] Failed to install WHM Configurations. Please install it manually." >&2;
           fi;
         ''
-        ]));
+      ]);
     };
   };
 }
